subscriber/requirements.txt
paho-mqtt
pymongo
py2neo
mysql-connector-python